Restructure and sqlite database insertion
