+++ imports/api/UserMethods.ts
@@ -12,7 +12,6 @@ export async function getUserByEmail(email: string): Promise<User> {
     return result[0];
 }
 
-export async function createUser(name: string, email: string, password: string): Promise<void> {
-    const hashedPassword = await Meteor.callAsync('server_createUser', { name, email, password });
-    return hashedPassword;
+export async function createUser(user: User): Promise<void> {
+    return await Meteor.callAsync('server_createUser', user);
 }
