+++ server/ServerBookMethods.ts
@@ -4,13 +4,6 @@ import Book from '/imports/types/book';
 
 // Define Meteor methods
 Meteor.methods({
-    // User methods
-    async server_getUsers(): Promise<any> {
-        return await executeQuery('SELECT * FROM Users');
-    },
-    async server_getUserByEmail(email: string): Promise<any> {
-        return await executeQuery('SELECT * FROM Users WHERE email = ?', [email]);
-    },
     // Book methods
     async server_getBooks(): Promise<any> {
         return await executeQuery('SELECT * FROM Books');
